support both wasm and common run
